rpi/requirements.txt
paho-mqtt==1.5.1
psutil==5.8.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0